# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "MailChampApi",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/Templates/{templateId}" : {
      "get" : {
        "tags" : [ "Templates" ],
        "summary" : "Gets the requested template.",
        "parameters" : [ {
          "name" : "templateId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Templates" : {
      "post" : {
        "tags" : [ "Templates" ],
        "summary" : "Creates a new template.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateInput"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateInput"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateInput"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Templates/{templateId}/name" : {
      "put" : {
        "tags" : [ "Templates" ],
        "summary" : "Updates the name of the selected template.",
        "parameters" : [ {
          "name" : "templateId",
          "in" : "path",
          "description" : "The template ID of the template to update.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "description" : "The template ID of the template to update.",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "The new template name.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateNameInput"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateNameInput"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateNameInput"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Templates/{templateId}/content" : {
      "put" : {
        "tags" : [ "Templates" ],
        "summary" : "Updates the content of the selected template.",
        "parameters" : [ {
          "name" : "templateId",
          "in" : "path",
          "description" : "The template ID of the template to update.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "description" : "The template ID of the template to update.",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "The new template content.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateContentInput"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateContentInput"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateContentInput"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Template"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ProblemDetails" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "nullable" : true
          },
          "title" : {
            "type" : "string",
            "nullable" : true
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "detail" : {
            "type" : "string",
            "nullable" : true
          },
          "instance" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : { }
      },
...
      "Template" : {
        "type" : "object",
        "properties" : {
          "templateId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "templateName" : {
            "type" : "string",
            "nullable" : true
          },
          "content" : {
            "type" : "string",
            "nullable" : true
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateModified" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "additionalProperties" : false,
        "example" : {
          "dateCreated" : "2000-01-23T04:56:07.000+00:00",
          "templateName" : "templateName",
          "dateModified" : "2000-01-23T04:56:07.000+00:00",
          "templateId" : 0,
          "content" : "content"
        }
      },
      "TemplateInput" : {
        "type" : "object",
        "properties" : {
          "templateName" : {
            "type" : "string",
            "nullable" : true
          },
          "content" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "TemplateNameInput" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "TemplateContentInput" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      }
    }
  }
}